Remove nodeId from PATCH requests
The new version of the API doesn't accept a nodeId in PATCH requests.
This might change in the future, but at the moment nodeId must be
removed for PATCH calls to succeed.

This commit also removes `enabled` from all calls, as it's an internal
API flag used for soft deletion and it shouldn't be updated by the user.