Correct test for externalId changes
The previous test was not really testing the bug reported in #20, as the
library compares the previous hash with the hash _of the object being
passed to the function_, so the only way to trigger the bug is to send
the full customer again, but with a different externalId. If you only
send the 'externalId', it will always be different from the previous
hash that contains a name and an email as well.